libptrace 1.0-rc2
Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>

Python 3.7 support.

Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>

Do not expose the avl tree on the public interface.

Better to have it removed before finalizing the 1.0 release.  It's better if
people do not start relying on avl tree internals.

pt_iterator_breakpoint_begin() has now been split up in:
  - pt_iterator_breakpoint_begin_process()
  - pt_iterator_breakpoint_begin_thread()

This is because on a process level are tracked in pt_process, and on a thread
level in pt_thread.  The rest of the iterator functions however remain the same
so there is no need to duplicate those at the moment.

Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>

Resolve issue #2 for the public header files.

This is the most urgent part of issue #2, as the public headers invite
fragility from external compilers.

Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>

Improve the installer.

- No longer use RMDir /r but selectively delete what was installed.
- Detect python versions installed on the system and offer to install bindings
  for those versions.
- Detect whether Python 2.7 is 32-bit or 64-bit, as this is not reflected in
  the registry.

Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>

Bugfixes

- Do not detect WoW64 processes when we've been built in 32-bit mode.  This is
  irrelevant and leads to errors.

- Fix a NULL deref when NtQueryObjectName returns a zero filled UNICODE_STRING.

Signed-off-by: Ronald Huizer <rhuizer@hexpedition.com>